requirements.txt
altgraph==0.17.2
lxml==4.7.1
macholib==1.15.2
overload==1.1
Pillow==9.0.0
pyinstaller==4.8
pyinstaller-hooks-contrib==2021.5
python-pptx==0.6.21
XlsxWriter==3.0.2
